argumentiation_analysis/requirements.txt
# requirements.txt

# Core Semantic Kernel library
semantic-kernel>=1.0.0,<2.0.0 
# Ou une version spécifique si vous en ciblez une, ex: semantic-kernel==1.5.1

# LLM Connectors (OpenAI/Azure) - SK les installe souvent, mais explicitons
openai>=1.0.0

# Configuration Environment
python-dotenv>=1.0.0

# JPype for Java/Tweety interaction
jpype1>=1.5.0 # Assurez-vous d'une version compatible avec votre Python/JDK

# UI Components
ipywidgets>=8.0.0
jupyter-ui-poll>=0.2.0 # For blocking UI cell execution
ipython # Implicitly needed for display in notebooks

# Data Handling (for InformalAnalysisPlugin)
pandas>=2.0.0

# HTTP Requests (for UI data fetching, CSV download)
requests>=2.30.0

# Cryptography (for UI config encryption)
cryptography>=42.0.0 # Utilise Fernet et PBKDF2

# Asyncio helper for Jupyter
nest-asyncio>=1.6.0

# --- Jupyter Kernel Support ---
ipykernel>=6.0.0 

